+++ main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"context"
 
-	"github.com/iron-io/functions/api/middlewares"
 	"github.com/iron-io/functions/api/server"
 )
 
@@ -12,7 +11,6 @@ func main() {
 
 	funcServer := server.NewFromEnv(ctx)
 	// Setup your custom extensions, listeners, etc here
-	middlewares.SetupJwtAuth(funcServer)
 
 	funcServer.Start(ctx)
 }
